+++ chunk-CDUAHQXN.js
@@ -1,0 +1,63 @@
+import {
+  AuthorisedAdminGuard
+} from "./chunk-Y7HC7MXS.js";
+import "./chunk-A5X5SB7F.js";
+import "./chunk-XI4ZLZAC.js";
+import "./chunk-OLCS3CIF.js";
+import "./chunk-RWTEO4RZ.js";
+import "./chunk-P7CK6ZAK.js";
+import "./chunk-5UGYVUD7.js";
+import "./chunk-BQISQ53F.js";
+import "./chunk-HT5GXKXQ.js";
+import "./chunk-TPDHL3PI.js";
+import "./chunk-QNBYRJPA.js";
+import "./chunk-X5UG4F5Q.js";
+import "./chunk-7A2HMJBQ.js";
+import "./chunk-ODRM6BWH.js";
+import "./chunk-62T6XHVB.js";
+import "./chunk-KWSTWQNB.js";
+
+// src/app/systems/systems.routes.ts
+var ROUTES = [
+  {
+    path: ":id",
+    loadComponent: () => import("./chunk-SHCY4R6G.js").then((m) => m.SystemsComponent),
+    children: [
+      {
+        path: "about",
+        loadComponent: () => import("./chunk-BOA4NEXW.js").then((m) => m.SystemAboutComponent)
+      },
+      {
+        path: "modules",
+        loadComponent: () => import("./chunk-2MNLD2ND.js").then((m) => m.SystemModulesComponent)
+      },
+      {
+        path: "triggers",
+        loadComponent: () => import("./chunk-LQPXUETJ.js").then((m) => m.SystemTriggersComponent)
+      },
+      {
+        path: "zones",
+        loadComponent: () => import("./chunk-FFMPCDQJ.js").then((m) => m.SystemZonesComponent)
+      },
+      {
+        path: "metadata",
+        loadComponent: () => import("./chunk-MYILQWM7.js").then((m) => m.SystemMetadataComponent)
+      },
+      {
+        path: "extend/:id",
+        loadComponent: () => import("./chunk-ETD4PF66.js").then((m) => m.ExtensionOutletComponent)
+      },
+      {
+        path: "history",
+        canActivate: [AuthorisedAdminGuard],
+        loadComponent: () => import("./chunk-RITUMYC4.js").then((m) => m.SettingsHistoryViewComponent)
+      },
+      { path: "**", redirectTo: "about" }
+    ]
+  },
+  { path: "**", redirectTo: "-" }
+];
+export {
+  ROUTES
+};
+//# sourceMappingURL=chunk-CDUAHQXN.js.map