+++ chunk-JIG5JSZV.js
@@ -1,0 +1,129 @@
+import {
+  AuthorisedAdminGuard
+} from "./chunk-Y7HC7MXS.js";
+import "./chunk-A5X5SB7F.js";
+import "./chunk-XI4ZLZAC.js";
+import "./chunk-OLCS3CIF.js";
+import "./chunk-RWTEO4RZ.js";
+import "./chunk-P7CK6ZAK.js";
+import "./chunk-5UGYVUD7.js";
+import "./chunk-BQISQ53F.js";
+import "./chunk-HT5GXKXQ.js";
+import "./chunk-TPDHL3PI.js";
+import "./chunk-QNBYRJPA.js";
+import "./chunk-X5UG4F5Q.js";
+import "./chunk-7A2HMJBQ.js";
+import "./chunk-ODRM6BWH.js";
+import "./chunk-62T6XHVB.js";
+import "./chunk-KWSTWQNB.js";
+
+// src/app/admin/admin.routes.ts
+var ROUTES = [
+  {
+    path: "",
+    loadComponent: () => import("./chunk-GUO6SBU5.js").then((m) => m.PlaceComponent),
+    canActivate: [AuthorisedAdminGuard],
+    children: [
+      {
+        path: "about",
+        loadComponent: () => import("./chunk-3JN5TRWV.js").then((m) => m.PlaceDetailsComponent)
+      },
+      {
+        path: "database",
+        loadComponent: () => import("./chunk-JY4VIZZI.js").then((m) => m.PlaceDatabaseDetailsComponent)
+      },
+      {
+        path: "clusters",
+        loadComponent: () => import("./chunk-KJKIVCU2.js").then((m) => m.PlaceClusterDetailsComponent)
+      },
+      {
+        path: "clusters/:id",
+        loadComponent: () => import("./chunk-QKA7CPYT.js").then((m) => m.PlaceClusterTaskListComponent)
+      },
+      {
+        path: "interfaces",
+        loadComponent: () => import("./chunk-Z5L3LSRL.js").then((m) => m.AdminInterfacesComponent)
+      },
+      {
+        path: "brokers",
+        loadComponent: () => import("./chunk-WPIVI4YM.js").then((m) => m.AdminBrokersComponent)
+      },
+      {
+        path: "edge",
+        loadComponent: () => import("./chunk-SEBF7JV2.js").then((m) => m.PlaceEdgeComponent)
+      },
+      {
+        path: "edge/:id",
+        loadComponent: () => import("./chunk-SEBF7JV2.js").then((m) => m.PlaceEdgeComponent)
+      },
+      {
+        path: "extensions",
+        loadComponent: () => import("./chunk-YUSRHOK2.js").then((m) => m.PlaceExtensionsComponent)
+      },
+      {
+        path: "staff-api",
+        loadComponent: () => import("./chunk-7W7TNS6Q.js").then((m) => m.PlaceStaffAPIComponent)
+      },
+      {
+        path: "resource-imports",
+        loadComponent: () => import("./chunk-2RUFCIPO.js").then((m) => m.ResourceImportsComponent)
+      },
+      {
+        path: "schemas",
+        loadComponent: () => import("./chunk-IZ6WLEUM.js").then((m) => m.AdminSchemasComponent)
+      },
+      {
+        path: "api-keys",
+        loadComponent: () => import("./chunk-SRNR62N5.js").then((m) => m.AdminAPIKeysComponent)
+      },
+      {
+        path: "upload-storage",
+        loadComponent: () => import("./chunk-VSJ7UGPQ.js").then((m) => m.StorageComponent)
+      },
+      {
+        path: "signage-ai",
+        loadComponent: () => import("./chunk-4SSH7DZP.js").then((m) => m.SignageAIComponent)
+      },
+      {
+        path: "upload-library",
+        loadComponent: () => import("./chunk-SPMVSNBN.js").then((m) => m.UploadLibraryComponent)
+      },
+      {
+        path: "mailing-list",
+        children: [
+          {
+            path: "",
+            loadComponent: () => import("./chunk-C7LRNXL3.js").then((m) => m.EmailTemplatesComponent)
+          },
+          {
+            path: "edit",
+            loadComponent: () => import("./chunk-QY5OM5KG.js").then((m) => m.EmailTemplateFormComponent)
+          },
+          {
+            path: "edit/:id",
+            loadComponent: () => import("./chunk-QY5OM5KG.js").then((m) => m.EmailTemplateFormComponent)
+          },
+          { path: "**", redirectTo: "" }
+        ]
+      },
+      {
+        path: "signage-plugins",
+        loadComponent: () => import("./chunk-2EBW4GMF.js").then((m) => m.AdminSignagePluginsComponent)
+      },
+      {
+        path: "build-jobs",
+        loadComponent: () => import("./chunk-MXLNLLLE.js").then((m) => m.PlaceBuildListComponent)
+      },
+      {
+        path: "extend/:id",
+        loadComponent: () => import("./chunk-ETD4PF66.js").then((m) => m.ExtensionOutletComponent)
+      },
+      { path: "**", redirectTo: "about" }
+    ]
+  },
+  { path: "**", redirectTo: "" }
+];
+export {
+  ROUTES
+};
+//# sourceMappingURL=chunk-JIG5JSZV.js.map