+++ chunk-JUISUIHR.js
@@ -1,0 +1,73 @@
+import {
+  AuthorisedUserGuard
+} from "./chunk-OLHNKJMR.js";
+import {
+  AuthorisedAdminGuard
+} from "./chunk-Y7HC7MXS.js";
+import "./chunk-A5X5SB7F.js";
+import "./chunk-XI4ZLZAC.js";
+import "./chunk-OLCS3CIF.js";
+import "./chunk-RWTEO4RZ.js";
+import "./chunk-P7CK6ZAK.js";
+import "./chunk-5UGYVUD7.js";
+import "./chunk-BQISQ53F.js";
+import "./chunk-HT5GXKXQ.js";
+import "./chunk-TPDHL3PI.js";
+import "./chunk-QNBYRJPA.js";
+import "./chunk-X5UG4F5Q.js";
+import "./chunk-7A2HMJBQ.js";
+import "./chunk-ODRM6BWH.js";
+import "./chunk-62T6XHVB.js";
+import "./chunk-KWSTWQNB.js";
+
+// src/app/zones/zones.routes.ts
+var ROUTES = [
+  {
+    path: ":id",
+    loadComponent: () => import("./chunk-B42W3CD6.js").then((m) => m.ZonesComponent),
+    children: [
+      {
+        path: "about",
+        loadComponent: () => import("./chunk-KX4RRH3U.js").then((m) => m.ZoneAboutComponent)
+      },
+      {
+        path: "systems",
+        loadComponent: () => import("./chunk-2VR7ZXDZ.js").then((m) => m.ZoneSystemsComponent)
+      },
+      {
+        path: "triggers",
+        canActivate: [AuthorisedUserGuard],
+        data: { role_only: true },
+        loadComponent: () => import("./chunk-2KUGYGZ5.js").then((m) => m.ZoneTriggersComponent)
+      },
+      {
+        path: "children",
+        loadComponent: () => import("./chunk-YDHP4UHI.js").then((m) => m.ZoneChildrenComponent)
+      },
+      {
+        path: "metadata",
+        loadComponent: () => import("./chunk-6HLLPFIQ.js").then((m) => m.ZoneMetadataComponent)
+      },
+      {
+        path: "groups",
+        canActivate: [AuthorisedAdminGuard],
+        loadComponent: () => import("./chunk-4I4LUKW4.js").then((m) => m.ZoneGroupsComponent)
+      },
+      {
+        path: "extend/:id",
+        loadComponent: () => import("./chunk-ETD4PF66.js").then((m) => m.ExtensionOutletComponent)
+      },
+      {
+        path: "history",
+        canActivate: [AuthorisedAdminGuard],
+        loadComponent: () => import("./chunk-RITUMYC4.js").then((m) => m.SettingsHistoryViewComponent)
+      },
+      { path: "**", redirectTo: "about" }
+    ]
+  },
+  { path: "**", redirectTo: "-" }
+];
+export {
+  ROUTES
+};
+//# sourceMappingURL=chunk-JUISUIHR.js.map