+++ chunk-KLZ36AWU.js
@@ -1,0 +1,37 @@
+import "./chunk-KWSTWQNB.js";
+
+// src/app/drivers/drivers.routes.ts
+var ROUTES = [
+  {
+    path: ":id",
+    loadComponent: () => import("./chunk-5WEYTRL4.js").then((m) => m.DriversComponent),
+    children: [
+      {
+        path: "about",
+        loadComponent: () => import("./chunk-OLEBSXUB.js").then((m) => m.DriverAboutComponent)
+      },
+      {
+        path: "docs",
+        loadComponent: () => import("./chunk-L3OPPPQH.js").then((m) => m.DriverDocsComponent)
+      },
+      {
+        path: "modules",
+        loadComponent: () => import("./chunk-AUEUTSFA.js").then((m) => m.DriverModulesComponent)
+      },
+      {
+        path: "extend/:id",
+        loadComponent: () => import("./chunk-ETD4PF66.js").then((m) => m.ExtensionOutletComponent)
+      },
+      {
+        path: "history",
+        loadComponent: () => import("./chunk-RITUMYC4.js").then((m) => m.SettingsHistoryViewComponent)
+      },
+      { path: "**", redirectTo: "about" }
+    ]
+  },
+  { path: "**", redirectTo: "-" }
+];
+export {
+  ROUTES
+};
+//# sourceMappingURL=chunk-KLZ36AWU.js.map