+++ chunk-QBQJ2MC6.js
@@ -1,0 +1,51 @@
+import {
+  AuthorisedAdminGuard
+} from "./chunk-Y7HC7MXS.js";
+import "./chunk-A5X5SB7F.js";
+import "./chunk-XI4ZLZAC.js";
+import "./chunk-OLCS3CIF.js";
+import "./chunk-RWTEO4RZ.js";
+import "./chunk-P7CK6ZAK.js";
+import "./chunk-5UGYVUD7.js";
+import "./chunk-BQISQ53F.js";
+import "./chunk-HT5GXKXQ.js";
+import "./chunk-TPDHL3PI.js";
+import "./chunk-QNBYRJPA.js";
+import "./chunk-X5UG4F5Q.js";
+import "./chunk-7A2HMJBQ.js";
+import "./chunk-ODRM6BWH.js";
+import "./chunk-62T6XHVB.js";
+import "./chunk-KWSTWQNB.js";
+
+// src/app/modules/modules.routes.ts
+var ROUTES = [
+  {
+    path: ":id",
+    loadComponent: () => import("./chunk-2M5UR25O.js").then((m) => m.ModulesComponent),
+    children: [
+      {
+        path: "about",
+        loadComponent: () => import("./chunk-7JKBGGTS.js").then((m) => m.ModuleAboutComponent)
+      },
+      {
+        path: "systems",
+        loadComponent: () => import("./chunk-BMEZP4EJ.js").then((m) => m.ModuleSystemsComponent)
+      },
+      {
+        path: "extend/:id",
+        loadComponent: () => import("./chunk-ETD4PF66.js").then((m) => m.ExtensionOutletComponent)
+      },
+      {
+        path: "history",
+        canActivate: [AuthorisedAdminGuard],
+        loadComponent: () => import("./chunk-RITUMYC4.js").then((m) => m.SettingsHistoryViewComponent)
+      },
+      { path: "**", redirectTo: "about" }
+    ]
+  },
+  { path: "**", redirectTo: "-" }
+];
+export {
+  ROUTES
+};
+//# sourceMappingURL=chunk-QBQJ2MC6.js.map