+++ chunk-ZIB6KU5F.js
@@ -1,0 +1,56 @@
+import {
+  AuthorisedAdminGuard
+} from "./chunk-Y7HC7MXS.js";
+import "./chunk-A5X5SB7F.js";
+import "./chunk-XI4ZLZAC.js";
+import "./chunk-OLCS3CIF.js";
+import "./chunk-RWTEO4RZ.js";
+import "./chunk-P7CK6ZAK.js";
+import "./chunk-5UGYVUD7.js";
+import "./chunk-BQISQ53F.js";
+import "./chunk-HT5GXKXQ.js";
+import "./chunk-TPDHL3PI.js";
+import "./chunk-QNBYRJPA.js";
+import "./chunk-X5UG4F5Q.js";
+import "./chunk-7A2HMJBQ.js";
+import "./chunk-ODRM6BWH.js";
+import "./chunk-62T6XHVB.js";
+import "./chunk-KWSTWQNB.js";
+
+// src/app/users/users.routes.ts
+var ROUTES = [
+  {
+    path: ":id",
+    loadComponent: () => import("./chunk-K66C2MPY.js").then((m) => m.UsersComponent),
+    children: [
+      {
+        path: "about",
+        loadComponent: () => import("./chunk-KVXJNNT2.js").then((m) => m.UserAboutComponent)
+      },
+      {
+        path: "metadata",
+        loadComponent: () => import("./chunk-B6DCTL3W.js").then((m) => m.UserMetadataComponent)
+      },
+      {
+        path: "groups",
+        canActivate: [AuthorisedAdminGuard],
+        loadComponent: () => import("./chunk-OCEZ7A7S.js").then((m) => m.UserGroupsComponent)
+      },
+      {
+        path: "history",
+        canActivate: [AuthorisedAdminGuard],
+        loadComponent: () => import("./chunk-IYHBWY2A.js").then((m) => m.UserHistoryComponent)
+      },
+      {
+        path: "extend/:id",
+        loadComponent: () => import("./chunk-ETD4PF66.js").then((m) => m.ExtensionOutletComponent)
+      },
+      { path: "**", redirectTo: "about" }
+    ]
+  },
+  { path: "**", redirectTo: "-" }
+];
+export {
+  ROUTES
+};
+//# sourceMappingURL=chunk-ZIB6KU5F.js.map